+++ pkg/tui/logging.go
@@ -1,0 +1,13 @@
+package tui
+
+import "log"
+
+var debugLogging = false
+
+// debug receives a string or strings, and logs it to the debug log if debugLogging is enabled
+func debug(msg ...string) {
+	if !debugLogging {
+		return
+	}
+	log.Printf("%s\n", msg)
+}
