+++ pkg/tui/tui.go
@@ -2,138 +2,20 @@ package tui
 
 import (
 	"fmt"
-	"log"
 	"time"
 
 	"github.com/PagerDuty/go-pagerduty"
-	"github.com/charmbracelet/bubbles/help"
 	"github.com/charmbracelet/bubbles/key"
 	"github.com/charmbracelet/bubbles/table"
-	"github.com/charmbracelet/bubbles/textinput"
-	"github.com/charmbracelet/bubbles/viewport"
 	tea "github.com/charmbracelet/bubbletea"
-	"github.com/clcollins/srepd/pkg/pd"
 )
 
-var debugLogging = false
-
-const waitTime = time.Millisecond * 1
-const defaultInputPrompt = " $ "
-
-func debug(msg ...string) {
-	if !debugLogging {
-		return
-	}
-	log.Printf("%s\n", msg)
-}
-
-func (m *model) setStatus(msg string) {
-	var d []string
-
-	m.status = fmt.Sprint(msg)
-
-	d = append(d, "setStatus")
-	d = append(d, msg)
-
-	log.Printf("%s\n", d)
-}
+const (
+	waitTime           = time.Millisecond * 1
+	defaultInputPrompt = " $ "
+)
 
 type errMsg struct{ error }
-
-type model struct {
-	err error
-
-	config   *pd.Config
-	editor   []string
-	launcher ClusterLauncher
-
-	table table.Model
-	input textinput.Model
-	// This is a hack since viewport.Model doesn't have a Focused() method
-	viewingIncident bool
-	incidentViewer  viewport.Model
-	help            help.Model
-
-	status string
-
-	incidentList           []pagerduty.Incident
-	selectedIncident       *pagerduty.Incident
-	selectedIncidentNotes  []pagerduty.IncidentNote
-	selectedIncidentAlerts []pagerduty.IncidentAlert
-
-	teamMode bool
-}
-
-func newTableWithStyles() table.Model {
-	debug("newTableWithStyles")
-	t := table.New(table.WithFocused(true))
-	t.SetStyles(tableStyle)
-	return t
-}
-
-func newTextInput() textinput.Model {
-	debug("newTextInput")
-	i := textinput.New()
-	i.Prompt = " $ "
-	i.CharLimit = 32
-	i.Width = 50
-	return i
-}
-
-func newHelp() help.Model {
-	debug("newHelp")
-	h := help.New()
-	h.ShowAll = false
-	return h
-}
-
-func newIncidentViewer() viewport.Model {
-	debug("newIncidentViewer")
-	vp := viewport.New(100, 100)
-	vp.Style = incidentViewerStyle
-	return vp
-}
-
-func InitialModel(
-	// "d" is to avoid a conflict with the "debug" function
-	d bool,
-	token string,
-	teams []string,
-	user string,
-	ignoredusers []string,
-	editor []string,
-	launcher ClusterLauncher,
-) (tea.Model, tea.Cmd) {
-	if d {
-		debugLogging = true
-	}
-
-	debug("InitialModel")
-	var err error
-
-	m := model{
-		editor:   editor,
-		launcher: launcher,
-		help:     newHelp(),
-		table:    newTableWithStyles(),
-		input:    newTextInput(),
-		// INCIDENTVIEWER
-		incidentViewer: newIncidentViewer(),
-		status:         "",
-	}
-
-	// This is an ugly way to handle this error
-	// We have to set the m.err here instead of how the errMsg is handled
-	// because the Init() occurs before the Update() and the errMsg is not
-	// preserved
-	pd, err := pd.NewConfig(token, teams, user, ignoredusers)
-	m.config = pd
-	m.err = err
-
-	return m, func() tea.Msg {
-		return errMsg{err}
-	}
-}
 
 func (m model) Init() tea.Cmd {
 	debug("Init")
